Allow for management of individual manager configurations via the API.
This doesn't actually expose the global config, which in fact serves
as a set of global defaults. Each manager can have an independent load
balancer connection as well as some number of hashed keys to pick services
to manager. To setup a load balancer only node, you would set the keys
to zero with a valid loadbalancer configuration. To setup a scale manager
only node, you would set the loadbalancer to none with an appropriate
number of non-zero keys.