Adding DNS loadbalancer and implicit API service.
We can now support multiple loadbalancers simultaneously. I've added
a simple (Nginx-like) loadbalancer module which wires everything through
dnsmasq. All services with the same URL will resolve to the same group
of IPs. This is useful for creating services that use DNS load balancing.

Additionally, there is now an implicit service which points to the API
servers in the backend. This service is available as the 'domain' specified
for pancake, so that requests on this domain will be routed to the API
servers themselves.