Loading of cloud/loadbalancer connections configurable
This moves away from having hardcoded, pre-defined, cloud
connection and loadbalancer connection classes. Now the system
can be configured to specify which connection class to load
and that class will be dynamically loaded.

A sensible default class composed of the loadbalancer name/
cloud type is used if the configuration does not set an
explicit class.

This will should help the addition of newer connection classes
as well as specifying mock classes for testing.