Refactoring the manager locking
There are two locks in the system: the manager lock and zookeeper
client lock. Since the manager and zookeeper do a lot of back and
forth communication there are potentials for deadlocks to occur.

The zookeeper locks are trigger around watches: when watches are
declared and when watches are fired. To prevent deadlocks, the
policy in the manager should be that it cannot hold the manager
lock while setting up / clearing a lock in zookeeper. This
prevents the dead lock where the manager is holding his lock,
waiting on the zookeeper client watch that is trying to grab the
manager lock.