Reactor Authentication Refactor
Currently authentication in the reactor admin site is performed by passing
an "auth_token" parameter in GETs and POSTs, leading to the display of
URLs containing passwords.

This commit refactors authentication in reactor via the following changes:
- Authentication and session management is now handled by pyramid's built-in
  modules
- Login/logout pages have been added to the web UI
- Assets (e.g. bootstrap) have been split off into a seperate handler so
  that they will properly render in unathenticated contexts (e.g. login)
- All auth_key URL passing has been stripped out
- The CLI has been updated to login for authenticated requests and use
  the session management cookie returned.

The purpose of this commit is to acheive correctness; features such as the
ability to change the admin password via the admin GUI will be added in later
commits.

Change-Id: Idb150e02eb53a92ec1d535333607c7025c2ce1e6