Reactor Active Directory Organazation Unit support
This patch adds support for an optional "orgunit" endpoint
configuration parameter. If present, the machine account will
be created in the given OU.

This patch also fixes up account deletion, which was
previously non-functional, and fixes a bug in port/socket
tracking caused by removal of a set entry while iterating
on the set.

Change-Id: I0a88fc5a2cf5d4e1e3337fa141e6749209db6456