Fix for RR-51 -- decommission instances become recomissioned and timeout.
As part of this fix, there are changes around how instances are manipulated.
This is not strictly necessary, but cleans the code. This came out of the fact
that I suspected the bug was caused by type confusion in instance IDs. Turns
out that the bug was simply that we expected the instance to re-confirm its
IP after being recomissioned.

Change-Id: I386c68b7b2028b33cfdc95842c7f915b7405ade6