partial fix for issue RR-77.
Python __path__ is set to the file from which a module is loaded.
Since the submodule is loaded from the parent module file, apparently
the __path__ symbol is set to the parent module's directory.  Workaround
is to do a path check and try again with a subdirectory if not found.
This is done just in case the __path__ symbol behaviour in python
changes in the future.

Change-Id: Ia6aab877bd918fd929cb144c84fce913d98a7c64