Sanity check new manager and endpoint names on the client side.
We weren't sanity checking new manager and endpoint names before
attempting to create a new instance. This meant if an empty or
malformed name was set, the resulting url for the instance would also
be invalid and the user was left with a blank screen when the
dashboard attempted to navigate to the config for the new instance.

Change-Id: I916cc7358eca6a986316f24c8fb29d5d0385783b