Fixups for cloud/loadbalancer module autoloading
At some point the cloud and loadbalancer modules
were moved into a subclass "connection". This patch
adjusts the autoloading accordingly, as well as
addresses other autoloading issues.

Next, this patch adds reporting of error messages
related to bad username/password/tenant config
values.

This patch also sets the service_type for
the nova client, as that seems to be required
for some clouds.

Change-Id: Ica82090970ae02e31d51a3dde2cef41d94469333