Adding feature to specify nics for VMs on osapi.
See RR-81.

The osapi cloud now optionally takes a network configuration parameter
networks for which interaces should be attached to the VM. If this
parameter is not configured, an interface is created for each
available network.

This change requires an relatively recent novaclient on the reactor
host. Without a sufficiently up to date novaclient, reactor falls back
to not presenting any networking options.

Change-Id: I7f0611874bd6499b51a3a436483c8d34df798b07