Loadbalancer fixes
This patch ups the min/max instance minimums to one - otherwise
loadbalancers are not able to properly report metrics, and no
scaling occurs. If the user wants zero instances they can stop
the service.

This patch also includes some code quality improvements in the
TCP loadbalancer, and adds connection management and proper
cleanup of socat processes and any bound sockets.

This patch removes the 'kill after use' parameter from the
tcp loadbalancer, since 1) it didn't work and 2) there's
no good way to implement it right now without more communication
between the endpoint class and the loadbalancer class.

Finally, this patch adds a new parameter, reconnect, to denote
the amount of time a client has to reconnect to an instance
before the instance is returned to the pool. While the instance
is in standby it is still counted towards the active instances,
so a rapid connect/disconnect/connect sequence will not cause
the scaling rule engine to get confused and do a
create/decommission/create.

Change-Id: I04ab031330737952b292dd564a29dab7ed8d5a4f