Fix reactor.config.ConfigSpec
reactor.config.ConfigSpec is a named tuple, formerly containing
an attribute "type". For pylint reasons, references to "type"
were changed to "typ", but the ConfigSpec declaration was not
updated, neither was the configuration parsing code.

This patch updates the ConfigSpec declaration and configuration
parsing code to use "typ" instead of "type".

Change-Id: I4fd4a61a7385b1bc444b79e129eae39fb2cd044c