Refactor of Zookeeper objects.
This patchset includes pylint cleanup across the codebase.
During the refactoring, I changed many of the method signatures,
and addressed many of the pylint complaints throughout the code.

It is highly likely that this has broken some pieces, but it
should be a more testable base on top of which we can build.

Change-Id: Ie9a8a2aafe8d4665a6f9edbfa6d6f43c891bf3c5