Allow the log to support arbitrary arguments.
This makes the log use the underlying zookeeper sequential
node feature. Now we can easily add arbitrary arguments to
the log for different events. Note that we still keep the
record_type structure (as it helps to formalize different
events).

This changeset also adds logging for the managers. This
will allow us to more easily debug potential problems from
the GUI, instead of having to drop into the server and check
log files.

Change-Id: Ib5599de878932422534852ad7dc4e8ad84822748