Force non-SSL LDAP to do creation and activation separately.
If you're using LDAP over a non-SSL connection, then it will silently
ignore activation attempts during machine creation. This is very difficult
to debug, because the creation calls will go through but the machine
accounts are not activated. If we're using non-SSL, simply force the
creation to proceed in multiple steps.

Change-Id: I1b585f73537f6ce5a35a55127fc7702bdcfc171d