Add pending count and restructure TCP loadbalancer.
This change adds a pending count that can be used for scaling
(e.g. pending < 1). It also restructures the TCP loadbalancer for
this purpose and to be more consistent and reliable with how locks
are given up.

Downside is that this changeset loses some test coverage from
the TCP loadbalancer. Will attempt to fix shortly.

Change-Id: Idbfbff5c4b356ba2800ef27083781e58edb04823