Disallow LDAP without SSL.
Okay, after a lot of poking around at SSL and eventual discover
of some authorative documentation, LDAP for AD does *not* support
changing the password for accounts over non-SSL connections.

Basically, this means that non-SSL connections are not useable for
our purposes and there's no point in supporting it. Administrators
will need to ensure that there is an appropriate certificate available
on all AD controllers.

Change-Id: I212753ca912e6430bf3a0d6728f5796cd05fa718