Refactor metrics() and reconnect logic.
There was the possibility that locks could be lost forever
previously. This was *really bad* as you would end up with
backends that are simply unreachable.

This refactors the logic (making it much simpler) and moves
the metrics collection right into the consumer. Currently,
metrics tests are dropped but could easily be recreated and
simplified (since this code is now much simpler).

Change-Id: I60b2f3f43d7ab6b0b3a0fe7523493585c482a09e