Allow endpoints to be deleted.
Because of the way trees are created recursively, endpoints occasionally
aren't deleted properly!
This is really tricky to fix properly (it'll require ripping out the
assumption that all Zookeeper trees are created recursively) but now for
we can workaround this issue by introduced a simple 'deleting' flag.
This will fire on all managers, and all clean up additional endpoint IPs.

Change-Id: Iab8b4a6929562103874b805ca5392f3de1650692