Enabling changing of template after creation.
This commit enables the user to change the template once an
endpoint has been created. This will preserve any configuration
changes made on-screen so far. The template is saved when the
configuration is saved, as would be expected.

This commit also changes the ordering used for the new endpoint
menu on the left, in order to be consistent with the ordering used
on the main page.

Change-Id: Iecbcd61579393991499c3afe79f0fb2aaa07b9c6