Fix a cpu chewing bug when the graphs are hidden (RR-99).
When the graphs aren't displaying, it seems like the CPU just
spins faster and faster and faster. That's because it does. Basically,
it seems like the resize() event is called constantly. But it calls
setTimeout() to schedule itself later. So effectively you just have
things piling on top of each other.

This has been fixed in the simplest way possible.

Change-Id: I423bfa19d9cc28c7ee0c08a3c5ecf078f6b65d3d