Add richer data, fix small issues in cloud/manager layer.
This commit addresses several small issues in the cloud and loadbalancer
connection layers. First and foremost, it adds a couple of extra args to
each (that could be useful for future cloud connections and lb connections
coming down the pipe). Secondly, it removes the custom "locks" tree for
load balancers, and replaces it with a generic and extensible object that
each loadbalancer/cloud can do with as it sees fit (this also required the
_cast_as function on the zookeeper objects).

This commit also fixes a couple of bugs in the config, discovered when
calling _manager_config() with no values, etc.

Finally, this commit adds some missing debug messages and tracebacks
where they are useful for diagnosing problems without adding much noise.

As you might have guessed, this commit is largely intended to prepare
the codebase for a new connection -- coming soon (RR-100).

Change-Id: I5ba7799cefdc128375dbd62a41dcc6a4629597bb