Restructure docs, setup and small fixes.
This commit began as a quest to implement a meta-reactor
service, that could start Reactors automatically. However,
this became quite convoluted (and of little utility).
It actually picked up several necessary restructuring fixes
to have the packages build on the different systems that
I was running it, and fixes to properly guess at the IP
address that was routeable. These are still valuable, and
part of the effort for RR-112.

Change-Id: I2015c0dfe26b35eac247e99e69d91dfe4252ebdd