Support rename operation.
As part of RR-112, I was writing documentation and really
required a rename operation for staging, upgrading, etc.
So I went and built it. Then it turned out that ths lead
to a much, much simpler way of handling deletes, so this
has allowed allow me to remove much of the convoluted
delete code!

Change-Id: I19fe18201d0a8ee3cf893fadedc482e84c037032