Enable floating IP support.
For real deployments, it's likely that users will require
floating IPs as a mechanism for distributing load (even if
it's only on Reactor itself). This adds it in a fairly simple
way, as it's very difficult to ensure HA services if you're
moving floating IPs. It really should be for non-dynamic
services (such as Reactor itself).

Part of the mega RR-109.

Change-Id: I0d11fa9a8728aedc703c78509a8628d91d8a1e87