Fix some unintended consequences of the rename change.
Managers now need to be aware of the fact that UUIDs and names
are two separate things. We carefully manage those mappings.
We use have to use those mappings for endpoint_ips(), which
thankfully is pretty clean.
As a result of having to do endpoint lookups for endpoint_ips()
the API now needs to use a cache for endpoint IP => name mappings.
Otherwise, it would need to constantly do a reverse lookup which
would be very expensive.

This is all part of RR-109.

Change-Id: Id46954013a007faebf4c9862ad5b63589f0de592