Associate was no working as expected.
You should be able to associate arbitrary instances. Unfortunately,
all list_instances() queries are filtered at the moment. This adds
a flag (that defaults to False) that allows for unfiltered queries.
Our list operations have been made much nicer thanks to the cache
anyways, so this allows for a mix of instances on a given endpoint.

Part of RR-109.

Change-Id: I3b8ab83e3cdb22365d16f5545ccc15906c1069c4