Break all reference loops by using weakref callbacks.
This dramatically simplifies how the objects are structured.
Instead of having to call break_refs() and unwatch() and objects,
we use weak references for callbacks in order to prevent loops
and make the structure much cleaner.

This is necessary to ensure that the process exits cleanly
(work for a future commit). Part of RR-109.

Change-Id: I1e17c52a8f111e65d7fefeef5745e680eda24203