Restructure common elements as AtomicRunnable.
This was a pattern that was oft-repeated through-out the code.
This commit refactors the code to reuse the common bits and have
less variation as a general rule.

This commit also applies a similar pattern to the ApiExtensions,
which previously relied on the __del__ operator.

Part of breaking apart server and client for RR-109.

Change-Id: I3743c140a77216718be4987ff80a595699eec565