Split client and server commands.
Split the client and server commands, both internally
as well as in packages. This includes simple scripts in
bin/ that will act when installed. Note that you can
still install via setup.py.

This includes tweaks to how commands are run in
setup-server.sh and the init scripts.

Part of RR-109.

Change-Id: I763628d8cfc0a071b96bded727c94acb77565c50