Splitting API and manager services, removing "cluster".
This caused a lot of problems for initial installations and
didn't add much value to the general case. We can instead handle
this externally (as I suspect most people will want anyways).

For simplicity, this splits all the CLI scripts into separate
files (ahhh... at last!) include the different "server" scripts.
As an additional benefit, this allows you to install any number
of reactor-managers on different hosts independently of the API
servers.

Part of making RR-112 work well out of the box.

Change-Id: I646886e7a7153041dd9b46638968450cdb4ecf17