RR-117 Add "kill-on-disconnect" to TCP LB
This patch adds a new setting, "disposable", to the
TCP loadbalancer, along with the plumbing required
for loadbalancer connections to discard VM instances.

When "disposable" is set, backend instances are
destroyed upon client disconnect. If a reconnect
timer is present, the backend instance is destroyed
when the reconnect timer expires.

Change-Id: Ic1acc558c6db881cc81bf03efe80d9ff7e448c57