Add minimum session time for KoD (wdil)
This patch adds a minimum session time for Kill-on-Disconnect.
Some (all?) RDP clients seem to connect to the RDP server
multiple times in rapid succession before establishing a connection
(possibly checking certificates, versions, etc). With KoD enabled,
this led to 2 VMs being prematurely killed for each established
connection.

With this patch, we add a minimum session time
(set to a sane value for real-world RDP deployments) which a
connection must satisfy before triggering KoD - otherwise,
the VM is returned to the VM pool.

Change-Id: I3188fe9557ad6eeff797e64d3bdb32b6f6412167